CELIX-438: Adds some additional api refactoring.
Also:
 - Adds support bundle group, which is used to cleanup the lb shell commands
 - Moves gen location for intermediate runtimes files

+++ libs/utils/include/celix_array_list.h
@@ -39,7 +39,7 @@ celix_array_list_t* celix_arrayList_createWithEquals(celix_arrayList_equals_fp e
 
 void celix_arrayList_destroy(celix_array_list_t *list);
 
-size_t celix_arrayList_size(const celix_array_list_t *list);
+int celix_arrayList_size(const celix_array_list_t *list);
 
 void* celix_arrayList_get(const celix_array_list_t *list, int index);
